requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
langchain>=0.2.0
langchain-community>=0.2.0
langchain-ollama>=0.1.0
langchain-openai>=0.1.0
langgraph>=0.1.0
chromadb>=0.5.0
sentence-transformers>=3.0.0
sqlalchemy>=2.0.0
streamlit>=1.35.0
pydantic>=2.7.0
python-dotenv>=1.0.0
slowapi>=0.1.9
requests>=2.32.0
